// chat-presence/app/src/generated/graphql.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: Date; output: Date; }
};

export type Contact = {
  __typename?: 'Contact';
  last_active_at: Scalars['Date']['output'];
  status: ContactStatus;
  user: Scalars['String']['output'];
};

export enum ContactStatus {
  Offline = 'OFFLINE',
  Online = 'ONLINE'
}